robo_tirilo/requirements.txt
pygame
supabase
google-generativeai
edge-tts
SpeechRecognition
pyaudio
gpiozero
python-dotenv
aiofiles
